% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_layout.R
\docType{data}
\name{data_layout_1}
\alias{data_layout_1}
\title{Datos layout_1 for conditional and marginal rates models}
\format{
Un data frame con 6347 rows y 13 columns.
}
\source{
data simulated from MSOAC parameters
}
\usage{
data_layout_1
}
\description{
Datos layout_1 for conditional and marginal rates models
}
\keyword{datasets}
