% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_revents.R
\docType{data}
\name{df_revents}
\alias{df_revents}
\title{Dato simulated}
\format{
Un data frame con 2601 rows y 11 columns
}
\source{
Simulated data
}
\usage{
df_revents
}
\description{
Data were simulated from clinical trials and its creation is explained in the simulated data vignette.
}
\keyword{datasets}
