% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Recurrent event simulated data}
\format{
A data frame with 22 rows and 11 variables:
\describe{
\item{Model}{Model name}
\item{paramcd}{Parameter name}
\item{overall_estimate}{Overall estimate}
\item{conf.int}{Confidence interval}
\item{conf_low}{Lower bound of confidence interval}
\item{conf_high}{Upper bound of confidence interval}
\item{p_value}{P-value}
\item{log_likelihood}{Log-likelihood}
\item{AIC}{Akaike Information Criterion}
\item{BIC}{Bayesian Information Criterion}
\item{type_model}{Type of model}
}
}
\source{
\"Simulated data based on the MSOAC dataset"
}
\usage{
simulated_data
}
\description{
Recurrent events in multiple sclerosis with repeated relapses simulated from the MSOAC dataset.
}
\keyword{datasets}
