% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_layout.R
\docType{data}
\name{data_layout_1}
\alias{data_layout_1}
\title{Datos layout_1 for conditional and marginal rates models}
\format{
Un data frame con 6347 rows y 13 columns.
}
\source{
data simulated
}
\usage{
data_layout_1
}
\description{
\item{ID}{participant ID}
\item{PARAMCD}{parameter}
\item{TSTART}{time study start}
\item{TSTOP}{time stop observation}
\item{TGAP}{time gap between events}
\item{STATUS}{status of the event}
\item{SEVENT}{sequence number of event per id}
\item{AGE}{age of the participant}
\item{SEX}{Sex of the participant}
\item{DISEASE_COURSE}{type of disease course}
\item{RACE}{race of the participant}
\item{TIME_SINCE_DIAGNOSIS}{time since diagnosis}#'
}
\keyword{datasets}
