% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_layout.R
\docType{data}
\name{data_layout_1}
\alias{data_layout_1}
\title{Data layout_1}
\format{
A data frame with 6347 rows and 13 columns:
\describe{
\item{ID}{participant ID}
\item{PARAMCD}{parameter}
\item{TSTART}{time study start}
\item{TSTOP}{time stop observation}
\item{TGAP}{time gap between events}
\item{STATUS}{status of the event}
\item{SEVENT}{sequence number of event per id}
\item{AGE}{age of the participant}
\item{SEX}{sex of the participant}
\item{DISEASE_COURSE}{type of disease course}
\item{RACE}{race of the participant}
\item{TIME_SINCE_DIAGNOSIS}{time since diagnosis}
}
}
\source{
Simulated data for conditional and marginal rates models
}
\usage{
data_layout_1
}
\description{
This dataset contains information about participants, including demographic details and event-related data.
}
\keyword{datasets}
