% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_layout.R
\docType{data}
\name{data_layout_2}
\alias{data_layout_2}
\title{Datos de ejemplo layout_2}
\format{
Un data frame con n filas y m columnas.
}
\source{
La fuente de los datos
}
\usage{
data_layout_2
}
\description{
Una descripción de los datos de ejemplo layout_2.
}
\keyword{datasets}
