% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_layout.R
\docType{data}
\name{data_layout_2}
\alias{data_layout_2}
\title{Data layout_2 for marginal hazard models}
\format{
a data frame with con 15765 rows y 12 columns.
}
\source{
data simulated from MSOAC parameters
}
\usage{
data_layout_2
}
\description{
Data layout_2 for marginal hazard models
}
\keyword{datasets}
