% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_layout.R
\docType{data}
\name{data_layout_3}
\alias{data_layout_3}
\title{Data layout_3 for count-based models}
\format{
Un data frame con 1313 rows y 9 columns.
}
\source{
data simulated from MSOAC parameters
}
\usage{
data_layout_3
}
\description{
Data layout_3 for count-based models
}
\keyword{datasets}
