% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_layout.R
\docType{data}
\name{data_layout_3}
\alias{data_layout_3}
\title{Data layout_3 for count-based models}
\format{
Un data frame con 1313 rows y 9 columns.
}
\source{
data simulated
}
\usage{
data_layout_3
}
\description{
\item{ID}{participant ID}
\item{PARAMCD}{parameter}
\item{COUNT}{number of confirmed events}
\item{SEVENT}{sequence number of event per id}
\item{AGE}{age of the participant}
\item{SEX}{Sex of the participant}
\item{DISEASE_COURSE}{type of disease course}
\item{RACE}{race of the participant}
\item{TIME_SINCE_DIAGNOSIS}{time since diagnosis}
}
\keyword{datasets}
