% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_revents.R
\docType{data}
\name{df_revents}
\alias{df_revents}
\title{Datos de ejemplo layout_1}
\format{
Un data frame con n filas y m columnas.
}
\source{
La fuente de los datos
}
\usage{
df_revents
}
\description{
Una descripción de los datos de ejemplo layout_1.
}
\keyword{datasets}
