% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_revents.R
\docType{data}
\name{df_revents}
\alias{df_revents}
\title{Dato simulated}
\format{
Un data frame con 2601 rows y 11 columns
}
\source{
MSOAC simulated data
}
\usage{
df_revents
}
\description{
This data was simulated from original MSOAC parameters and it is explained in vignette of simulated data
}
\keyword{datasets}
